move xarray dataarray creation functions into hi.utils.py move associated tests into correcte test file
