+++ imap_processing/hi/utils.py
@@ -1,6 +1,13 @@
 """IMAP-Hi utils functions."""
 
+from collections.abc import Sequence
 from enum import IntEnum
+from typing import Optional, Union
+
+import numpy as np
+import xarray as xr
+
+from imap_processing.cdf.imap_cdf_manager import ImapCdfAttributes
 
 
 class HIAPID(IntEnum):
@@ -25,3 +32,94 @@ class HIAPID(IntEnum):
             "45sensor" or "90sensor".
         """
         return self.name[1:3] + "sensor"
+
+
+def full_dataarray(
+    name: str,
+    attrs: dict,
+    coords: Optional[dict[str, xr.DataArray]] = None,
+    shape: Optional[Union[int, Sequence[int]]] = None,
+) -> xr.DataArray:
+    """
+    Generate an empty xarray.DataArray with appropriate attributes.
+
+    Data in DataArray are filled with FILLVAL defined in attributes
+    retrieved from ATTR_MGR with shape matching coordinates defined by
+    dims or overridden by optional `shape` input.
+
+    Parameters
+    ----------
+    name : str
+        Variable name.
+    attrs : dict
+        CDF variable attributes. Usually retrieved from ImapCdfAttributes.
+    coords : dict
+        Coordinate variables for the Dataset.
+    shape : int or tuple
+        Shape of ndarray data array to instantiate in the xarray.DataArray.
+
+    Returns
+    -------
+    data_array : xarray.DataArray
+        Meeting input specifications.
+    """
+    _attrs = attrs.copy()
+    dtype = attrs.pop("dtype", None)
+
+    # extract dims keyword argument from DEPEND_i attributes
+    dims = [v for k, v in sorted(attrs.items()) if k.startswith("DEPEND")]
+    # define shape of the ndarray to generate
+    if shape is None:
+        shape = [coords[k].data.size for k in dims]  # type: ignore
+    if hasattr(shape, "__len__") and len(shape) > len(dims):
+        dims.append("")
+
+    data_array = xr.DataArray(
+        np.full(shape, attrs["FILLVAL"], dtype=dtype),
+        name=name,
+        dims=dims,
+        attrs=attrs,
+    )
+    return data_array
+
+
+def create_dataset_variables(
+    variable_names: list[str],
+    variable_shape: Union[int, Sequence[int]],
+    att_manager_lookup_str: str = "{0}",
+) -> dict[str, xr.DataArray]:
+    """
+    Instantiate new `xarray.DataArray` variables.
+
+    Variable attributes are retrieved from CdfAttributeManager.
+
+    Parameters
+    ----------
+    variable_names : list[str]
+        List of variable names to create.
+    variable_shape : tuple[int]
+        Shape of the new variables data ndarray.
+    att_manager_lookup_str : str
+        String defining how to build the string passed to the
+        CdfAttributeManager in order to retrieve the CdfAttributes for each
+        variable. The string passed to CdfAttributeManager will be the result
+        of calling the `str.format()` method on this input string with the
+        variable name from `variable_names` as the single argument. Defaults to
+        "{0}".
+
+    Returns
+    -------
+    new_variables : dict[str, xarray.DataArray]
+        Dictionary of new xarray.DataArray variables.
+    """
+    attr_mgr = ImapCdfAttributes()
+    attr_mgr.add_instrument_global_attrs("hi")
+    attr_mgr.load_variable_attributes("imap_hi_variable_attrs.yaml")
+
+    new_variables = dict()
+    for var in variable_names:
+        attrs = attr_mgr.get_variable_attributes(
+            att_manager_lookup_str.format(var), check_schema=False
+        )
+        new_variables[var] = full_dataarray(var, attrs, shape=variable_shape)
+    return new_variables
